chore: merge cli and web
