+++ package.json
@@ -9,8 +9,8 @@
   ],
   "scripts": {
     "dev": "yarn workspace web dev",
-    "cli:upload": "yarn workspace @instagram-scraper/cli build && yarn workspace @instagram-scraper/cli start upload",
-    "cli:scrape": "yarn workspace @instagram-scraper/cli build && HEADLESS=false yarn workspace @instagram-scraper/cli start",
+    "cli:scrape": "HEADLESS=false yarn workspace web cli scrape",
+    "cli:upload": "yarn workspace web cli upload",
     "lint": "biome check ./ --write",
     "typecheck": "turbo typecheck",
     "prepare": "husky"
