--- packages/cli/.env.example
@@ -1,5 +1,0 @@
-INSTAGRAM_USERNAME=your_instagram_username
-INSTAGRAM_PASSWORD=your_instagram_password
-
-HEADLESS=true
-WEB_API=http://localhost:3000
