--- packages/cli/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@instagram-scraper/cli",
-  "version": "1.0.0",
-  "description": "CLI for Instagram saved posts scraper",
-  "main": "dist/index.js",
-  "bin": {
-    "instagram-scraper": "./dist/index.js"
-  },
-  "scripts": {
-    "build": "tsc",
-    "typecheck": "tsc --noEmit",
-    "start": "node dist/index.js",
-    "dev": "yarn build && node dist/index.js",
-    "clean": "rm -rf dist"
-  },
-  "keywords": [
-    "instagram",
-    "scraper",
-    "cli"
-  ],
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "@internal/shared": "workspace:*",
-    "@playwright/test": "1.55.0",
-    "@types/node-fetch": "2",
-    "dotenv": "^17.2.1",
-    "node-fetch": "2",
-    "zod": "^4.1.3"
-  },
-  "devDependencies": {
-    "@types/node": "^24.3.0",
-    "typescript": "^5.9.2"
-  },
-  "engines": {
-    "node": ">=22.0.0"
-  }
-}
