--- packages/shared/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "@internal/shared",
-  "version": "0.0.0",
-  "private": true
-}
