+++ packages/web/.env.example
@@ -7,3 +7,9 @@ AWS_REGION=
 AWS_ENDPOINT_URL=
 NEXT_PUBLIC_AWS_BUCKET_PROXY_URL=
 SOCKS5_PROXY=
+
+INSTAGRAM_USERNAME=your_instagram_username
+INSTAGRAM_PASSWORD=your_instagram_password
+
+HEADLESS=true
+WEB_API=http://localhost:3000
