+++ packages/web/package.json
@@ -23,7 +23,6 @@
   "dependencies": {
     "@aws-sdk/client-s3": "^3.879.0",
     "@ewnd9/turbo-scripts": "^0.2.3",
-    "@internal/shared": "workspace:*",
     "@t3-oss/env-nextjs": "^0.12.0",
     "@tanstack/react-query": "^5.69.0",
     "@trpc/client": "^11.0.0",
@@ -36,6 +35,7 @@
     "drizzle-kit": "^0.30.5",
     "drizzle-orm": "^0.41.0",
     "next": "^15.2.3",
+    "node-fetch": "^3.3.2",
     "pg-boss": "^10.3.2",
     "postgres": "^3.4.4",
     "react": "^19.0.0",
@@ -48,12 +48,14 @@
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
+    "@playwright/test": "1.55.0",
     "@swc-node/register": "^1.11.1",
     "@swc/core": "^1.13.5",
     "@tailwindcss/postcss": "^4.0.15",
     "@types/node": "^24.3.0",
     "@types/react": "^19.0.0",
     "@types/react-dom": "^19.0.0",
+    "dotenv": "^17.2.2",
     "postcss": "^8.5.3",
     "tailwindcss": "^4.0.15",
     "tsx": "^4.19.0",
